Move design to core:presentation
