+++ core/presentation/design/src/main/AndroidManifest.xml
